app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
ezdxf==1.3.4
pydantic==2.9.0
python-multipart==0.0.9
matplotlib==3.9.1
shapely==2.0.5
# AI providers (optional – install only the one you use)
openai==1.43.0
groq==0.11.0
python-dotenv==1.0.1